requirements.txt
pandas
imagehash
opencv-contrib-python
Pillow
colorthief